TE-7587 Moved TransferTypeValidatorTrait to Shared
