+++ src/Spryker/Shared/Transfer/TypeValidation/TransferTypeValidatorTrait.php
@@ -5,7 +5,7 @@
  * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
-namespace Spryker\Zed\Transfer\Business\Transfer\TypeValidation;
+namespace Spryker\Shared\Transfer\TypeValidation;
 
 use InvalidArgumentException;
