+++ src/Spryker/Zed/Transfer/Business/Model/Generator/ClassDefinition.php
@@ -9,10 +9,10 @@ namespace Spryker\Zed\Transfer\Business\Model\Generator;
 
 use ArrayObject;
 use Spryker\DecimalObject\Decimal;
+use Spryker\Shared\Transfer\TypeValidation\TransferTypeValidatorTrait;
 use Spryker\Zed\Transfer\Business\Exception\InvalidAssociativeTypeException;
 use Spryker\Zed\Transfer\Business\Exception\InvalidAssociativeValueException;
 use Spryker\Zed\Transfer\Business\Exception\InvalidNameException;
-use Spryker\Zed\Transfer\Business\Transfer\TypeValidation\TransferTypeValidatorTrait;
 use Spryker\Zed\Transfer\TransferConfig;
 use Zend\Filter\Word\CamelCaseToUnderscore;
 use Zend\Filter\Word\UnderscoreToCamelCase;
